pyproject.toml
[tool.poetry]
name = "lambda-runtime-tweeter-lambda"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
python-twitter = "^3.5"
jeffy = "^0.1.3"

[tool.poetry.dev-dependencies]
aws-sam-cli = "^0.40.0"
black = "^19.10b0"
isort = "^4.3.21"
ipython = "^7.12.0"
boto3 = "^1.11.12"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
